requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
pyttsx3==2.90
ultralytics==8.0.196
timm==0.9.7
python-socketio==5.9.0
eventlet==0.33.3
pytesseract pillow